Move GRPC2 types to concordium-base.
This is so that they may be used by the client.
