+++ concordium-consensus/package.yaml
@@ -10,15 +10,12 @@ copyright:           "Concordium"
 extra-source-files:
 - README.md
 - ChangeLog.md
-- ../concordium-grpc-api/v2/concordium/types.proto
-
 
 custom-setup:
   dependencies:
     - base  >= 4.7
     - Cabal >= 2.2
     - directory >= 1.3
-    - proto-lens-setup >= 0.4
 
 # Metadata used when publishing your package
 # synopsis:            Short description of your package
@@ -61,8 +58,6 @@ dependencies:
 - lmdb
 - primitive
 - proto-lens >= 0.7
-- proto-lens-protobuf-types >= 0.7
-- proto-lens-runtime >= 0.7
 
 default-extensions:
 - FlexibleContexts
@@ -78,7 +73,7 @@ default-extensions:
 flags:
   dynamic:
     manual: True
-    default: True
+    default: False
 
 library:
   source-dirs: src
@@ -105,11 +100,6 @@ library:
               ghc-options: -shared -dynamic
             else:
               ghc-options: -static
-
-
-  generated-exposed-modules:
-    Proto.Concordium.Types
-    Proto.Concordium.Types_Fields
 
 # Other options might be needed, such has -dynamic -lHSrts or -lHSrts-ghc8.4.3
 # Possibly this depends on the platform
